+++ uniCloud-aliyun/cloudfunctions/router/service/admin/system/app/sys/update.js
@@ -23,7 +23,26 @@ module.exports = {
 			appid,
 			type,
 			name,
-			description
+			description,
+			// 1.13.4 新增以下字段（主要用于生成统一发布页和APP升级中心）
+			enable_upgrade_center,
+			introduction,
+			icon_url,
+			screenshot,
+			app_android,
+			app_ios,
+			mp_weixin,
+			mp_alipay,
+			mp_baidu,
+			mp_toutiao,
+			mp_qq,
+			mp_kuaishou,
+			mp_lark,
+			mp_jd,
+			mp_dingtalk,
+			quickapp,
+			h5,
+			store_list,
 		} = data;
 		// 参数合法校验开始-----------------------------------------------------------
 		let formRulesRes = await formRules.update(event);
@@ -50,7 +69,26 @@ module.exports = {
 				appid,
 				type,
 				name,
-				description
+				description,
+				// 1.13.4 新增以下字段（主要用于生成统一发布页和APP升级中心）
+				enable_upgrade_center,
+				introduction,
+				icon_url,
+				screenshot,
+				app_android,
+				app_ios,
+				mp_weixin,
+				mp_alipay,
+				mp_baidu,
+				mp_toutiao,
+				mp_qq,
+				mp_kuaishou,
+				mp_lark,
+				mp_jd,
+				mp_dingtalk,
+				quickapp,
+				h5,
+				store_list,
 			}
 		});
 		// 业务逻辑结束-----------------------------------------------------------