+++ uni_modules/vk-unicloud/vk_modules/vk-unicloud-page/libs/vk-unicloud/vk-unicloud-callFunctionUtil.js
@@ -899,9 +899,9 @@ class CallFunctionUtil {
 	getFileSuffix(obj = {}) {
 		let {
 			file,
-			filePath
+			filePath,
+			suffix = "png"
 		} = obj;
-		let suffix = "png";
 		if (filePath) {
 			let suffixName = filePath.substring(filePath.lastIndexOf(".") + 1);
 			if (suffixName && suffixName.length < 5) suffix = suffixName;