+++ uni_modules/vk-unicloud/vk_modules/vk-unicloud-page/libs/vk-unicloud/vk-unicloud-user-center.js
@@ -1016,4 +1016,16 @@ export default {
 			url: 'user/kh/unbindQQ',
 		});
 	},
+	/**
+	 * App升级中心 - 检测是当前版本是否需要升级
+	 * res 返回参数说明
+	 * @param {Number} code 错误码，0表示成功
+	 * @param {String} msg 详细信息
+	 */
+	checkVersion(obj = {}) {
+		return callFunction({
+			...obj,
+			url: 'user/pub/checkVersion',
+		});
+	},
 };